* tizen_3.0:
  sensord: add dummy api library for tv profile
  sensord: accumulate pedometer data
  sensord: version up 2.0.9
  sensord: fix incorrect return type
  sensor: enable pedometer sensor
  sensor: fix return value from false to 0
  sensord: add pre_start() that should be executed before starting regardless of clients
  sensord: change log level from error to warning for the warning case
  sensor: change the hard coded path for supporting 64bit platform
  Set SmackProcessLabel to System
  sensord: store unknown rotation event when auto_rotation sensor starts
  Fix for build errors of packages that are using deprecated sensor api

+++ src/server/command_worker.cpp
@@ -91,7 +91,7 @@ void command_worker::init_cmd_handlers(void)
 	m_cmd_handlers[CMD_FLUSH]				= &command_worker::cmd_flush;
 }
 
-int command_worker::create_sensor_raw_list(int client_perms, std::vector<raw_data_t *> &raw_list)
+size_t command_worker::create_sensor_raw_list(int client_perms, std::vector<raw_data_t *> &raw_list)
 {
 	size_t total_raw_data_size = 0;
 	vector<sensor_base *> sensors;
